requirements.txt
flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.0
numpy==1.26.0
scikit-learn==1.5.0
# Note: Install en_core_web_sm model with: python -m spacy download en_core_web_sm 
